refactor: create interface for MutableMultiSet
